# Clean up and parallelize testsuite