--- a/tests/opt/.gitignore
+++ b/tests/opt/.gitignore
@@ -1,1 +1,2 @@
 *.log
+run-test.mk