--- a/tests/sat/.gitignore
+++ b/tests/sat/.gitignore
@@ -1,1 +1,2 @@
 *.log
+run-test.mk