- Factor out creating a Makefile from the `*.ys`/`*.v`/`*.sh` in the working directory to reduce redundancy in `run-test.sh` files
- Convert several tests to use Makefiles supporting `-j` for concurrency

Here are some crude timing benchmarks. First, on my i7-3520M (quad core) laptop:

.               | `make -j4 test` (before) | `make -j4 test` (after) | `make -j1`
--------------- | ------------------------ | ----------------------- | ----------
Wall time       | 899 s                    | 544 s                   | 1070 s
usr+sys time    | 1253 s                   | 1515 s                  | 1069 s
= concurrency   | 1.39                     | 2.78                    | 1.00

Resulting in a concurrency boost of **+100%** and a runtime change of **-39%**. I'm pretty sure the increase in total processor time is due to (increased) thermal throttling, so the measured effect on runtime is probably lower than it could be.

On my i7-4790 (8-core) desktop:

.               | `make -j8 test` (before) | `make -j8 test` (after) | `make -j1`
--------------- | ------------------------ | ----------------------- | ----------
Wall time       | 648 s                    | 328 s                   | 796 s
usr+sys time    | 975 s                    | 1282 s                  | 797 s
= concurrency   | 1.50                     | 3.91                    | 1.00

Which is a concurrency boost of **+161%** and a runtime change of **-49%**.