@clairexen 2020-09-17
I'm not a big fan of the use of `nullglob` and `failglob` and `pipefail` and so forth, as they make it harder for people to understand the scripts bc they are changing the default behavior of bash, so looking at a random line of the script might mislead people bc they lack the information that the default behavior of bash has been switched to something else in a different place in the script, or even another script included via `source`.