@Xiretza 2020-09-17
Fair reasoning in principle, but `failglob` in particular is basically just free sanity checking, since it's exceedingly rare that one actually wants to silently expand a glob pattern to itself if there's no match. The `pipefail` is pure habit, I'll remove that. As for the `nullglob` in `tests/memories/run-test.sh`, I copied that line from `tests/simple/run-test.sh`, if it's not desired I can remove it from both.