@Xiretza 2020-09-21
Removed all occurences of `nullglob`/`failglob`/`pipefail`.